Add code that enables solution of draft task
Draft task has 2 data sets: train and test. After task is printed on
screen, the user is given a shell to work out the solution on train data
set. When solution is ready, user exits the shell. The last command
entered by user in that shell is considered the solution. The last
command is then applied to the test data set and the result is compared
against the known correct result to determine whether user's solution is
correct.