Solution is checked by string comparison against known correct variants
Parsing user solution to check whether it's correct is prohibitive.

Treating user solution as a black box command, then re-running it with a
hidden data set to check correctness is complex. Also black box
approach is not suitable, b/c we don't need _any_ solution that produces
correct results. Instead, for educational purposes for a particular task
we need particular solution, e.g. one that uses "cut" and not "awk". So
we need to somehow peek into solution, but not parse.

Thus instead of parsing we simply prepare a list of known correct
solutions and search user's solution among them. Similar to how Duolingo
does this in written exercises.